Move tests into their app
