+++ .github/workflows/backend-test.yml
@@ -22,12 +22,11 @@ jobs:
       - name: Install codecov
         working-directory: backend
         run: pip install codecov
-
       - name: pytest
         working-directory: backend
         run: |
             set +e
-            OUTPUT=$(poetry run py.test --cov-report xml --cov=. --cov-fail-under 100)
+            OUTPUT=$(poetry run pytest --cov-report xml --cov=. --cov-fail-under 100)
             STATUS=$?
             echo "$OUTPUT"
             codecov
@@ -48,9 +47,8 @@ jobs:
       - name: Install dependencies
         working-directory: backend
         run: poetry install
-
       - name: pytest
         working-directory: backend
-        run: poetry run py.test integration_tests/
+        run: poetry run pytest --integration
         env:
           STRIPE_SECRET_KEY: ${{ secrets.STRIPE_SECRET_KEY }}
