+++ backend/conferences/tests/test_models.py
