+++ backend/conferences/tests/test_schema.py
