--- backend/integration_tests/conftest.py
@@ -1,1 +1,0 @@
-from tests.conftest import *  # noqa
