+++ backend/payments/tests/test_stripe_buy_conference_ticket.py
@@ -1,4 +1,4 @@
-from unittest.mock import patch
+from unittest.mock import Mock, patch
 
 import stripe
 from django.conf import settings
@@ -13,7 +13,59 @@ stripe.api_key = settings.STRIPE_SECRET_KEY
 
 
 @mark.django_db
-def test_buy_conference_ticket(graphql_client, user, ticket_fare_factory):
+def test_success_order_correctly_generates_a_client_secret(
+    graphql_client, user, ticket_fare_factory
+):
+    graphql_client.force_login(user)
+
+    fare = ticket_fare_factory()
+
+    with patch("payments.providers.stripe.stripe.PaymentIntent.create") as create:
+        create_mock = Mock()
+        create_mock.id = "123456789"
+        create_mock.client_secret = "abc"
+        create.return_value = create_mock
+
+        response = graphql_client.query(
+            """mutation ($conference: ID!, $items: [CartItem!]!) {
+                buyTicketWithStripe(input: {
+                    conference: $conference,
+                    items: $items
+                }) {
+                    __typename
+
+                    ... on StripeClientSecret {
+                        clientSecret
+                    }
+                }
+            }""",
+            variables={
+                "conference": fare.conference.code,
+                "items": [{"id": fare.id, "quantity": 1}],
+            },
+        )
+
+    assert "errors" not in response, response["errors"]
+
+    assert response["data"]["buyTicketWithStripe"]["__typename"] == "StripeClientSecret"
+    assert response["data"]["buyTicketWithStripe"]["clientSecret"] == "abc"
+
+    order = Order.objects.first()
+
+    assert order.items.count() == 1
+
+    item = order.items.all()[0]
+
+    assert item.quantity == 1
+    assert item.item_object == fare
+    assert item.unit_price == fare.price
+
+
+@mark.django_db
+@mark.integration
+def test_stripe_integration_buying_conference_ticket(
+    graphql_client, user, ticket_fare_factory
+):
     graphql_client.force_login(user)
 
     fare = ticket_fare_factory()
