+++ backend/voting/tests/factories/vote.py
@@ -1,8 +1,8 @@
 import factory.fuzzy
 from factory.django import DjangoModelFactory
 from pytest_factoryboy import register
-from tests.submissions.factories import SubmissionFactory
-from tests.users.factories import UserFactory
+from submissions.tests.factories import SubmissionFactory
+from users.tests.factories import UserFactory
 from voting.models import Vote
